backend/requirements.txt
# backend/requirements.txt

Django>=5.2
djangorestframework>=3.14
django-cors-headers>=3.15
djangorestframework-simplejwt>=5.3
whitenoise>=6.6
python-dotenv>=1.0
channels>=4.0
channels-redis>=4.2
google-auth
google-api-python-client
google-oauthlib
